+++ tests/ingest/reducers/ltdsphinxtechnote_test.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 import datetime
 from pathlib import Path
 
+import structlog
 import yaml
 
 from ook.ingest.reducers.ltdsphinxtechnote import ReducedLtdSphinxTechnote
@@ -12,6 +13,8 @@ from ook.ingest.reducers.ltdsphinxtechnote import ReducedLtdSphinxTechnote
 
 def test_sqr035_reduction() -> None:
     """Test ReducedLtdSphinxTechnote using SQR-035 as a test dataset."""
+    logger = structlog.get_logger("ook")
+
     sqr035data = (
         Path(__file__).parent.parent.parent / "data" / "content" / "sqr-035"
     )
@@ -20,7 +23,7 @@ def test_sqr035_reduction() -> None:
     url = "https://sqr-035.lsst.io/"
 
     reduced_technote = ReducedLtdSphinxTechnote(
-        html_source=html_source, url=url, metadata=metadata
+        html_source=html_source, url=url, metadata=metadata, logger=logger
     )
 
     assert reduced_technote.url == "https://sqr-035.lsst.io/"
@@ -52,58 +55,58 @@ def test_sqr035_reduction() -> None:
     ]
 
     sections = reduced_technote.sections
-    assert sections[0].url == "https://sqr-035.lsst.io/#context"
-    assert sections[0].headers == [reduced_technote.h1, "1   Context"]
-    assert sections[0].content.startswith(
+    assert sections[1].url == "https://sqr-035.lsst.io/#context"
+    assert sections[1].headers == [reduced_technote.h1, "1   Context"]
+    assert sections[1].content.startswith(
         "This document does three things: lays out the elements and practice "
         "we use for kubernetes-based services, outlines guidelines for best "
         "practices, and a discussion of current and upcoming technological "
         "choices for implementation."
     )
 
-    assert sections[1].url == "https://sqr-035.lsst.io/#docker-image-release"
-    assert sections[1].headers == [
+    assert sections[2].url == "https://sqr-035.lsst.io/#docker-image-release"
+    assert sections[2].headers == [
         reduced_technote.h1,
         "2   Elements",
         "2.1   Docker Image Release",
     ]
     assert (
-        sections[2].url == "https://sqr-035.lsst.io/#configuration-management"
+        sections[3].url == "https://sqr-035.lsst.io/#configuration-management"
     )
-    assert sections[2].headers == [
+    assert sections[3].headers == [
         reduced_technote.h1,
         "2   Elements",
         "2.2   Configuration Management",
     ]
 
-    assert sections[3].url == "https://sqr-035.lsst.io/#secrets"
-    assert sections[3].headers == [
+    assert sections[4].url == "https://sqr-035.lsst.io/#secrets"
+    assert sections[4].headers == [
         reduced_technote.h1,
         "2   Elements",
         "2.3   Secrets",
     ]
 
     assert (
-        sections[4].url == "https://sqr-035.lsst.io/#deployment-orchestration"
+        sections[5].url == "https://sqr-035.lsst.io/#deployment-orchestration"
     )
-    assert sections[4].headers == [
+    assert sections[5].headers == [
         reduced_technote.h1,
         "2   Elements",
         "2.4   Deployment Orchestration",
     ]
 
-    assert sections[5].url == "https://sqr-035.lsst.io/#configuration-control"
-    assert sections[5].headers == [
+    assert sections[6].url == "https://sqr-035.lsst.io/#configuration-control"
+    assert sections[6].headers == [
         reduced_technote.h1,
         "2   Elements",
         "2.5   Configuration Control",
     ]
 
-    assert sections[6].url == "https://sqr-035.lsst.io/#elements"
-    assert sections[6].headers == [reduced_technote.h1, "2   Elements"]
+    assert sections[7].url == "https://sqr-035.lsst.io/#elements"
+    assert sections[7].headers == [reduced_technote.h1, "2   Elements"]
 
-    assert sections[7].url == "https://sqr-035.lsst.io/#deployment-add-ons"
-    assert sections[7].headers == [
+    assert sections[8].url == "https://sqr-035.lsst.io/#deployment-add-ons"
+    assert sections[8].headers == [
         reduced_technote.h1,
         "3   Deployment add-ons",
     ]
@@ -115,6 +118,7 @@ def test_dmtn021_reduction() -> None:
     This document contained an HtmlComment element that was tripping up
     iter_sphinx_sections. This test proves that we've handled it.
     """
+    logger = structlog.get_logger("ook")
     data_root = (
         Path(__file__).parent.parent.parent / "data" / "content" / "dmtn-021"
     )
@@ -123,9 +127,27 @@ def test_dmtn021_reduction() -> None:
     url = "https://dmtn-021.lsst.io/"
 
     reduced_technote = ReducedLtdSphinxTechnote(
-        html_source=html_source, url=url, metadata=metadata
+        html_source=html_source, url=url, metadata=metadata, logger=logger
     )
 
     assert reduced_technote.h1 == (
         "Implementation of Image Difference Decorrelation"
     )
+
+
+def test_dmtn139_reduction() -> None:
+    """Test with DMTN-139, which does not have any sectioning content."""
+    logger = structlog.get_logger("ook")
+    data_root = (
+        Path(__file__).parent.parent.parent / "data" / "content" / "dmtn-139"
+    )
+    html_source = (data_root / "index.html").read_text()
+    metadata = yaml.safe_load((data_root / "metadata.yaml").read_text())
+    url = "https://dmtn-139.lsst.io/"
+
+    reduced_technote = ReducedLtdSphinxTechnote(
+        html_source=html_source, url=url, metadata=metadata, logger=logger
+    )
+
+    assert reduced_technote.h1 == ("LSST Image Service Architecture")
+    assert len(reduced_technote.sections) == 1