+++ CHANGELOG.rst
@@ -2,6 +2,22 @@
 Change log
 ##########
 
+0.3.0 (2020-07-17)
+==================
+
+- Improved ingest reliability:
+
+  - For Lander_ (PDF) content, added a heuristic that rejects TeX that Pandoc might let through.
+  - Handle AASTeX technotes that don't have full Lander_ site.
+    Specifically, AASTeX technotes don't include the handle in the TeX source, so instead we use the document's URL.
+  - If a Lander_ site doesn't have an abstract, we fall back to using the first content chunk.
+  - Support Lander_ docs without content
+  - Support Sphinx technotes that include content before the first subsection header.
+
+- Improved logging during ingests, including logging of records when an insertion into the Algolia index fails.
+
+- Refreshed all pinned dependencies
+
 0.2.0 (2020-07-02)
 ==================